ai-engineer/03-industry-frameworks/llamaindex-pdf-rag/python/requirements.txt
llama-index-core
llama-index-llms-ollama
llama-index-embeddings-huggingface
llama-index-readers-file
pymupdf
